Extract builder methods from `PartialCycle`
